implement transaction modules with snapshot

+++ src/migrations/001-create-profile-table.ts
@@ -3,7 +3,7 @@ import { pool } from "../config/database";
 export async function up() {
   await pool.query(`
     CREATE TABLE profile (
-      id INT AUTO_INCREMENT PRIMARY KEY,
+      id BIGINT AUTO_INCREMENT PRIMARY KEY,
       email VARCHAR(255) NOT NULL UNIQUE,
       password VARCHAR(255) NOT NULL,
       first_name VARCHAR(255) NOT NULL,
